DataEgress code updated to read the decrypt flag from DynamoDB and sk… (#16)
* DataEgress code updated to read the decrypt flag from DynamoDB and skip the decryption if it is set to false.
Two versions of save functions made to save encrypted and decrypted data on S3. Encrypted data is written to destination bucket with the Metadata containing IV, CipherText and KeyID for end user to consume. Tests added and code refactored to skip the compression for already encrypted data coming from HTME
* Fixup: Format Python code with Black

Co-authored-by: DataWorks-CI <48289519+DataWorks-CI@users.noreply.github.com>